- Separated the BigDecimal and BigInteger implementations - Prepared the project to be completely multiplatform ready - Updated the publishing names - Updated the README
