+++ bigdecimal/build.gradle
@@ -1,0 +1,5 @@
+kotlin.sourceSets {
+    commonMain.dependencies {
+        api project(":biginteger")
+    }
+}
