+++ gradle/common.gradle
@@ -7,3 +7,5 @@ kotlin.sourceSets {
         api "org.jetbrains.kotlin:kotlin-test-annotations-common:$kotlin_version"
     }
 }
+
+project.ext.set("commonStructure", true)
