+++ gradle/darwin.gradle
@@ -1,0 +1,31 @@
+apply from: rootProject.file("gradle/ide.gradle")
+apply from: rootProject.file("gradle/native.gradle")
+
+kotlin {
+    targets {
+        if (project.ext.ideaActive) {
+            fromPreset(project.ext.ideaPreset, 'darwin')
+        } else {
+            fromPreset(presets.iosArm64, 'iosArm64')
+            fromPreset(presets.iosArm32, 'iosArm32')
+            fromPreset(presets.iosX64, 'iosX64')
+            fromPreset(presets.macosX64, 'macosX64')
+        }
+    }
+    sourceSets {
+        darwinMain { dependsOn commonMain }
+        darwinTest
+
+        if (!project.ext.ideaActive) {
+            configure([iosArm32Main, iosArm64Main, iosX64Main, macosX64Main]) {
+                dependsOn darwinMain
+            }
+
+            configure([iosArm32Test, iosArm64Test, iosX64Test, macosX64Test]) {
+                dependsOn darwinTest
+            }
+        }
+    }
+}
+
+project.ext.set("hasNative", true)
