+++ gradle/publish.gradle
@@ -1,3 +1,4 @@
+apply plugin: 'maven'
 apply plugin: 'maven-publish'
 
 apply from: project.rootProject.file('gradle/pom.gradle')
@@ -13,27 +14,55 @@ task stubDokka(type: Jar) {
     classifier 'kdoc'
 }
 
-publishing {
-    repositories {
-        maven {
-            if (publishLocal) {
-                url globalM2
-            } else {
-                def user = 'komputing'
-                def repo = 'kbignumbers'
-                def name = 'kbignumbers'
-                url = "https://api.bintray.com/maven/$user/$repo/$name/;publish=0"
-                credentials {
-                    username = project.hasProperty('bintrayUser') ? project.property('bintrayUser') : System.getenv('BINTRAY_USER')
-                    password = project.hasProperty('bintrayApiKey') ? project.property('bintrayApiKey') : System.getenv('BINTRAY_API_KEY')
-                }
-                
-            }
-        }
-        maven { name = "testLocal"; url = "$rootProject.buildDir/m2" }
+def selectArtifactId(project, type, defaultName) {
+    def name = project.name
+    def hasNative = project.ext.has("hasNative") && project.ext.hasNative
+
+    switch (type) {
+        case 'metadata':
+            return "$name-common"
+            break
+        case 'kotlinMultiplatform':
+            if (!hasNative) return "$name-kotlinMultiplatform"
+            return "$name-native"
+            break
+        case 'jvm':
+            return "$name-jvm"
+            break
+        default:
+            return hasNative ? defaultName : "$name"
+            break
     }
+}
+
+def isAvailableForPublication(publication) {
+    def name = publication.name
+    def result = false
+    result = result | (name in ['jvm', 'js', 'metadata', 'kotlinMultiplatform'])
+    result = result | (project.ext.isLinuxHost && name == 'linuxX64')
+    result = result | (project.ext.isWinHost && name == 'mingwX64')
+    result = result | (project.ext.isMacosHost && name in ['iosX64', 'iosArm32', 'iosArm64', 'macosX64'])
+
+    return result
+}
+
+tasks.withType(AbstractPublishToMaven).all {
+    onlyIf { isAvailableForPublication(publication) }
+}
+
+publishing {
     publications.all {
         pom.withXml(configureMavenCentralMetadata)
+
+        def type = it.name
+        def id = selectArtifactId(project, type, it.artifactId)
+        it.artifactId = id
+
+        if (id.endsWith('native')) {
+            it.artifact stubJavadoc
+            it.artifact stubDokka
+            it.artifact emptyJar
+        }
     }
     kotlin.targets.all { target ->
         def publication = publishing.publications.findByName(target.name)
